% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_ratio.R
\name{best_contrast}
\alias{best_contrast}
\title{Find highest contrast color}
\usage{
best_contrast(x, y = c("#010101", "#FFFFFF"))
}
\arguments{
\item{x}{A vector of colors as described in \code{col} of \code{\link[=color]{color()}}. Must not
contain any \code{NA}.}

\item{y}{A vector of colors as described in \code{col} of \code{\link[=color]{color()}}. Must not
contain any \code{NA}.}
}
\value{
A vector of the same length as \code{x} with, for each element of \code{x}, the
element of \code{y} that has the highest contrast to \code{x}.
}
\description{
\code{best_contrast()} finds the color in \code{y} with the highest contrast to the
color \code{x}.
}
\examples{
best_contrast("red")
best_contrast("grey20")
best_contrast("white")

best_contrast(rainbow(10), rainbow(3))
}
