% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-blindness.R
\name{check_color_blindness}
\alias{check_color_blindness}
\title{Visualize color vision deficiency}
\usage{
check_color_blindness(col)
}
\arguments{
\item{col}{A \code{colors} object (see \code{\link[=color]{color()}}) or a vector of any of the three
kinds of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}
}
\value{
Invisibly \code{col}.
}
\description{
\code{check_color_blindness()} will showcase the effect of the three kinds of
color vision deficiency, Deuteranopia, Protanopia, and Tritanopia, at the
same time side by side in a plot.
}
\examples{
check_color_blindness(rainbow(10))

check_color_blindness(terrain.colors(10))
}
