% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyscale.R
\name{clr_grayscale}
\alias{clr_grayscale}
\alias{clr_greyscale}
\title{Transform colors to grayscale}
\source{
\url{https://tannerhelland.com/3643/grayscale-image-algorithm-vb6/}

\url{https://en.wikipedia.org/wiki/Luma}
}
\usage{
clr_grayscale(
  col,
  method = c("luma", "averaging", "min_decomp", "max_decomp", "red_channel",
    "green_channel", "blue_channel")
)

clr_greyscale(
  col,
  method = c("luma", "averaging", "min_decomp", "max_decomp", "red_channel",
    "green_channel", "blue_channel")
)
}
\arguments{
\item{col}{A \code{colors} object (see \code{\link[=color]{color()}}) or a vector of any of the three
kinds of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}

\item{method}{A character string specifying the grayscaling method. Can be
one of "luma", "averaging", "min_decomp", "max_decomp", "red_channel",
"green_channel" and "blue_channel". Defaults to "luma".}
}
\value{
A \code{colors} object of the same length as \code{col}.
}
\description{
\code{clr_grayscale()} has a selection of different methods to turn colors into
grayscale.
}
\details{
If \code{method = "averaging"} then the red, green and blue will be
averaged together to create the grey value. This method does a poor job of
representing the way the human eye sees color. If \code{method = "luma"} (the
default), a weighted average is used to calculate the grayscale values. The
BT. 709 method from the ITU Radiocommunication Sector have determined the
weights. If \code{method} is \code{"min_decomp"} or \code{"max_decomp"}, then a decomposition
method is used where the minimum or maximum color value have been selected
for the color value. So the color rgb(60, 120, 40) would have the "min_decomp"
value of 40 and "max_decomp" value of 120. If method is \code{"red_channel"},
\code{"green_channel"} or \code{"blue_channel"}, then the corresponding color channel
will be selected for the values of grayscale.
}
\examples{

plot(clr_grayscale(rainbow(10)))

plot(clr_grayscale(terrain.colors(10)))

viridis_colors <- c(
  "#4B0055FF", "#422C70FF", "#185086FF", "#007094FF",
  "#008E98FF", "#00A890FF", "#00BE7DFF", "#6CD05EFF",
  "#BBDD38FF", "#FDE333FF"
)

plot(clr_grayscale(viridis_colors, method = "luma"))
plot(clr_grayscale(viridis_colors, method = "averaging"))
plot(clr_grayscale(viridis_colors, method = "min_decomp"))
plot(clr_grayscale(viridis_colors, method = "max_decomp"))
plot(clr_grayscale(viridis_colors, method = "red_channel"))
plot(clr_grayscale(viridis_colors, method = "green_channel"))
plot(clr_grayscale(viridis_colors, method = "blue_channel"))
}
