% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negate.R
\name{clr_negate}
\alias{clr_negate}
\title{Negate colors in RGB space}
\usage{
clr_negate(col)
}
\arguments{
\item{col}{A \code{colors} object (see \code{\link[=color]{color()}}) or a vector of any of the three
kinds of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}
}
\value{
A \code{colors} object of the same length as \code{col}.
}
\description{
Negate colors in RGB space
}
\details{
The negation of color is happening in the red-green-blue colorspace
RGB. This means if we take the specification for orange which is
rgb(255, 165, 0), then we negate by taking the opposite number on the scale
from 0 to 255, leaving us with rgb(0, 90, 255), which is a shade of blue.
}
\examples{
clr_negate("orange")

terr <- color(terrain.colors(10))

terr
clr_negate(terr)

plot(terr)
plot(clr_negate(terr))
}
