% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{clr_rotate}
\alias{clr_rotate}
\title{Rotate the colors around the hue wheel}
\source{
\url{https://en.wikipedia.org/wiki/HCL_color_space}
}
\usage{
clr_rotate(col, degrees = 0)
}
\arguments{
\item{col}{A \code{colors} object (see \code{\link[=color]{color()}}) or a vector of any of the three
kinds of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}

\item{degrees}{A number between 0 and 360, denoting the amount of degrees the
colors should be rotated. Defaults to 0.}
}
\value{
A \code{colors} object of the same length as \code{col}.
}
\description{
Rotate the colors around the hue wheel
}
\details{
The colors will be transformed to HCL color space
(Hue-Chroma-Luminance) where the hue of the color will be rotated.
}
\examples{
plot(clr_rotate(terrain.colors(10)))

plot(clr_rotate(terrain.colors(10), degrees = 90))

plot(clr_rotate(terrain.colors(10), degrees = 180))

plot(clr_rotate(rep("magenta", 11), degrees = seq(0, 360, length.out = 11)))
}
