% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{Turn vector to color vector}
\usage{
color(col)
}
\arguments{
\item{col}{a color object or vector of any of the three kinds of R color
specifications, i.e., either a color name (as listed by colors()), a
hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a
positive integer i meaning palette()[i].}
}
\value{
a color object.
}
\description{
Turn vector to color vector
}
\examples{
terrain_10 <- color(terrain.colors(15))

terrain_10

plot(terrain_10)
}
