% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_ratio.R
\name{contrast_ratio}
\alias{contrast_ratio}
\title{Contrast ratio between colors}
\source{
\url{https://www.w3.org/TR/UNDERSTANDING-WCAG20/visual-audio-contrast-contrast.html}
}
\usage{
contrast_ratio(x, y)
}
\arguments{
\item{x}{A length 1 color object (see \code{\link[=color]{color()}}) or a length 1 vector of any
of the three kinds of R color specifications, i.e., either a color name (as
listed by \code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or
a positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}

\item{y}{A color object (see \code{\link[=color]{color()}}) or a vector of any of the three kinds
of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}
}
\value{
A numerical vector of the same length as \code{y} of the calculated
contrast ratios.
}
\description{
\code{contrast_ratio()} calculates the contrast ratio between the color \code{x} and
the color(s) \code{y}. Contrast ratios can range from 1 to 21 with 1 being no
contrast (i.e., same color) and 21 being highest contrast.
}
\details{
The formula used for calculating a contrast ratio between two colors is

\deqn{(L1 + 0.05) / (L2 + 0.05)}

where

\itemize{
\item L1 is the relative luminance of the lighter of the colors, and
\item L2 is the relative luminance of the darker of the colors.
}

Relative luminance is calculated according to
\url{https://www.w3.org/TR/2008/REC-WCAG20-20081211/#relativeluminancedef}.
}
\examples{
contrast_ratio("red", "blue")
contrast_ratio("grey20", grey.colors(10))
contrast_ratio("white", c("white", "black"))
}
