% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{clr_extract_chroma}
\alias{clr_extract_chroma}
\title{Extract HCL components}
\usage{
clr_extract_chroma(col)
}
\arguments{
\item{col}{A \code{colors} object (see \code{\link[=color]{color()}}) or a vector of any of the three
kinds of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}
}
\value{
Numeric vector of values.
}
\description{
Extract the hue, chroma, or luminance color components from a vector of
colors.
}
\details{
The range of the value are:
\itemize{
\item Hue is ranging from 0 to 360.
\item Luminance is ranging from 0 to 100.
\item Chroma, while dependent on hue and luminance, will roughly be within 0 and
180.
}

Use \code{\link[=clr_extract]{clr_extract()}} if you are planning to extraction multiple components.
}
\examples{
clr_extract_hue(rainbow(100), "HCL")
clr_extract_chroma(rainbow(100))
clr_extract_luminance(rainbow(100))
}
\seealso{
Other Extraction: 
\code{\link{clr_extract}()},
\code{\link{clr_extract_hue}()},
\code{\link{extract_rgba}()}
}
\concept{Extraction}
