% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{is_color}
\alias{is_color}
\title{Test if object is a \code{colors} object}
\usage{
is_color(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{colors} class, else \code{FALSE}.
}
\description{
Test if object is a \code{colors} object
}
