% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prismatic-package.R
\docType{package}
\name{prismatic-package}
\alias{prismatic}
\alias{prismatic-package}
\title{prismatic: Color Manipulation Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Manipulate and visualize colors in a intuitive, low-dependency and functional way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilhvitfeldt.github.io/prismatic/}
  \item \url{https://github.com/EmilHvitfeldt/prismatic}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/prismatic/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

}
\keyword{internal}
