% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_ratio.R
\name{best_contrast}
\alias{best_contrast}
\title{Find highest contrast color}
\usage{
best_contrast(x, y = c("#010101", "#FFFFFF"))
}
\arguments{
\item{x}{Multiple colors}

\item{y}{Multiple colors}
}
\value{
The elements of `y` with highest contrast to `x`.
}
\description{
Finds the color in `y` with the highest contrast to the color `x`.
}
\examples{
best_contrast("red")
best_contrast("grey20")
best_contrast("white")

best_contrast(rainbow(10), rainbow(3))
}
