% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-blindness.R
\name{check_color_blindness}
\alias{check_color_blindness}
\title{Visualize color vision deficiency}
\usage{
check_color_blindness(col)
}
\arguments{
\item{col}{a color object or vector of any of the three kinds of R color
specifications, i.e., either a color name (as listed by colors()), a
hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a
positive integer i meaning palette()[i].

This function will showcase the effect of all 3 kinds of color vision
deficiency at the same time side by side.}
}
\value{
Nothing
}
\description{
Visualize color vision deficiency
}
\examples{
check_color_blindness(rainbow(10))

check_color_blindness(terrain.colors(10))
}
