% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyscale.R
\name{clr_grayscale}
\alias{clr_grayscale}
\title{Transform colors to greyscale}
\usage{
clr_grayscale(col)
}
\arguments{
\item{col}{a color object or vector of any of the three kinds of R color
specifications, i.e., either a color name (as listed by colors()), a
hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a
positive integer i meaning palette()[i].}
}
\value{
a color object of same length as col.
}
\description{
Transform colors to greyscale
}
\examples{

plot(clr_grayscale(rainbow(10)))

plot(clr_grayscale(terrain.colors(10)))

plot(clr_grayscale(hcl.colors(10, palette = "viridis")))
}
