% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\alias{colour}
\title{Turn vector of colors to \code{color} vector}
\usage{
color(col)

colour(col)
}
\arguments{
\item{col}{A \code{colors} object (see \code{\link[=color]{color()}}) or a vector of any of the three
kinds of R color specifications, i.e., either a color name (as listed by
\code{\link[grDevices:colors]{grDevices::colors()}}), a hexadecimal string (see \code{\link[=col2rgb]{col2rgb()}}), or a
positive integer \code{i} meaning \code{\link[grDevices:palette]{grDevices::palette()}}\verb{[i]}.}
}
\value{
A \code{colors} object of the same length as \code{col}. Returns hex 8 digits
form "#rrggbbaa". See \emph{Details}.
}
\description{
Turn vector of colors to \code{color} vector
}
\details{
Alpha values will be automatically added to hexcodes. If no alpha
value is present in \code{col}, it will default to no alpha (FF).
}
\examples{
terrain_10 <- color(terrain.colors(10))

terrain_10[1:4]

plot(terrain_10)

plot(terrain_10, labels = TRUE)

grey_10 <- color(gray.colors(10, start = 0, end = 1))

grey_10

plot(grey_10, labels = TRUE)
}
