% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\alias{colour}
\title{Turn vector to color vector}
\usage{
color(col)

colour(col)
}
\arguments{
\item{col}{a color object or vector of any of the three kinds of R color
specifications, i.e., either a color name (as listed by colors()), a
hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a
positive integer i meaning palette()[i].}
}
\value{
a colors object.
}
\description{
Turn vector to color vector
}
\details{
Alpha values will be automatically  added to hexcodes. If none at
present it will default to no alpha (FF).
}
\examples{
terrain_10 <- color(terrain.colors(10))

terrain_10[1:4]

plot(terrain_10)

plot(terrain_10, labels = TRUE)

plot(color(gray.colors(10)), labels = TRUE)
}
