% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-blindness.R
\name{clr_protan}
\alias{clr_protan}
\alias{clr_deutan}
\alias{clr_tritan}
\title{Simulate color vision deficiency}
\usage{
clr_protan(col, severity = 1)

clr_deutan(col, severity = 1)

clr_tritan(col, severity = 1)
}
\arguments{
\item{col}{a color object or vector of any of the three kinds of R color
specifications, i.e., either a color name (as listed by colors()), a
hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a
positive integer i meaning palette()[i].}

\item{severity}{A numeric, Severity of the color vision defect, a number
between 0 and 1. 0 means no deficiency, 1 means complete deficiency. Defaults
 to 1.}
}
\value{
a color object of same length as col.
}
\description{
Simulate color vision deficiency
}
\examples{
rainbow_colors <- color(rainbow(10))

plot(clr_protan(rainbow_colors))
plot(clr_deutan(rainbow_colors))
plot(clr_tritan(rainbow_colors))

viridis_colors <- color(hcl.colors(10, palette = "viridis"))

plot(clr_protan(viridis_colors))
plot(clr_deutan(viridis_colors))
plot(clr_tritan(viridis_colors))
}
