% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{is_color}
\alias{is_color}
\title{Test if the object is a color}
\usage{
is_color(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object inherits from the color class.
}
\description{
Test if the object is a color
}
